requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alphagenome==0.2.0
anndata==0.12.2
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.10.0
array-api-compat==1.12.0
attrs==25.3.0
biomcp-python==0.6.9
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
crc32c==2.7.1
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
donfig==0.8.1.post1
fonttools==4.60.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.38.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
immutabledict==4.2.1
iniconfig==2.1.0
intervaltree==3.1.0
jaxtyping==0.3.2
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
legacy-api-wrap==1.4.1
markdown-it-py==4.0.0
matplotlib==3.10.6
mcp==1.14.1
mdurl==0.1.2
ml_dtypes==0.5.3
multidict==6.6.4
natsort==8.4.0
numcodecs==0.16.3
numpy==2.3.3
openai==1.108.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scipy==1.16.2
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.0.2
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typeguard==4.4.4
typer==0.19.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.36.0
wadler_lindig==0.1.7
websockets==15.0.1
yarl==1.20.1
zarr==3.1.3
zstandard==0.25.0
